Refactor connectors
