+++ mcp_use/connectors/stdio.py
@@ -6,13 +6,11 @@ through the standard input/output streams.
 """
 
 import sys
-from typing import Any
 
 from mcp import ClientSession, StdioServerParameters
-from mcp.types import CallToolResult, Tool
 
 from ..logging import logger
-from ..task_managers import ConnectionManager, StdioConnectionManager
+from ..task_managers import StdioConnectionManager
 from .base import BaseConnector
 
 
@@ -39,14 +37,11 @@ class StdioConnector(BaseConnector):
             env: Optional environment variables.
             errlog: Stream to write error output to.
         """
+        super().__init__()
         self.command = command
         self.args = args or []  # Ensure args is never None
         self.env = env
         self.errlog = errlog
-        self.client: ClientSession | None = None
-        self._connection_manager: ConnectionManager | None = None
-        self._tools: list[Tool] | None = None
-        self._connected = False
 
     async def connect(self) -> None:
         """Establish a connection to the MCP implementation."""
@@ -81,108 +76,3 @@ class StdioConnector(BaseConnector):
 
             # Re-raise the original exception
             raise
-
-    async def disconnect(self) -> None:
-        """Close the connection to the MCP implementation."""
-        if not self._connected:
-            logger.debug("Not connected to MCP implementation")
-            return
-
-        logger.info("Disconnecting from MCP implementation")
-        await self._cleanup_resources()
-        self._connected = False
-        logger.info("Disconnected from MCP implementation")
-
-    async def _cleanup_resources(self) -> None:
-        """Clean up all resources associated with this connector."""
-        errors = []
-
-        # First close the client session
-        if self.client:
-            try:
-                logger.debug("Closing client session")
-                await self.client.__aexit__(None, None, None)
-            except Exception as e:
-                error_msg = f"Error closing client session: {e}"
-                logger.warning(error_msg)
-                errors.append(error_msg)
-            finally:
-                self.client = None
-
-        # Then stop the connection manager
-        if self._connection_manager:
-            try:
-                logger.debug("Stopping connection manager")
-                await self._connection_manager.stop()
-            except Exception as e:
-                error_msg = f"Error stopping connection manager: {e}"
-                logger.warning(error_msg)
-                errors.append(error_msg)
-            finally:
-                self._connection_manager = None
-
-        # Reset tools
-        self._tools = None
-
-        if errors:
-            logger.warning(f"Encountered {len(errors)} errors during resource cleanup")
-
-    async def initialize(self) -> dict[str, Any]:
-        """Initialize the MCP session and return session information."""
-        if not self.client:
-            raise RuntimeError("MCP client is not connected")
-
-        logger.info("Initializing MCP session")
-
-        # Initialize the session
-        result = await self.client.initialize()
-
-        # Get available tools
-        tools_result = await self.client.list_tools()
-        self._tools = tools_result.tools
-
-        logger.info(f"MCP session initialized with {len(self._tools)} tools")
-
-        return result
-
-    @property
-    def tools(self) -> list[Tool]:
-        """Get the list of available tools."""
-        if not self._tools:
-            raise RuntimeError("MCP client is not initialized")
-        return self._tools
-
-    async def call_tool(self, name: str, arguments: dict[str, Any]) -> CallToolResult:
-        """Call an MCP tool with the given arguments."""
-        if not self.client:
-            raise RuntimeError("MCP client is not connected")
-
-        logger.debug(f"Calling tool '{name}' with arguments: {arguments}")
-        result = await self.client.call_tool(name, arguments)
-        return result
-
-    async def list_resources(self) -> list[dict[str, Any]]:
-        """List all available resources from the MCP implementation."""
-        if not self.client:
-            raise RuntimeError("MCP client is not connected")
-
-        logger.debug("Listing resources")
-        resources = await self.client.list_resources()
-        return resources
-
-    async def read_resource(self, uri: str) -> tuple[bytes, str]:
-        """Read a resource by URI."""
-        if not self.client:
-            raise RuntimeError("MCP client is not connected")
-
-        logger.debug(f"Reading resource: {uri}")
-        resource = await self.client.read_resource(uri)
-        return resource.content, resource.mimeType
-
-    async def request(self, method: str, params: dict[str, Any] | None = None) -> Any:
-        """Send a raw request to the MCP implementation."""
-        if not self.client:
-            raise RuntimeError("MCP client is not connected")
-
-        logger.debug(f"Sending request: {method} with params: {params}")
-        return await self.client.request({"method": method, "params": params or {}})
